request y response conductor

+++ src/main/java/com/cootramixtol/sipd/entities/Conductor.java
@@ -29,7 +29,7 @@ public class Conductor {
 	@Column(name = "codigo")
 	private long codigo;
 	@Column(name = "identificacion", unique = true)
-	private int identificacion;
+	private long identificacion;
 	@Column(name = "nombres")
 	private String nombres;
 	@Column(name = "apellidos")
